Collection Areas, Gameplay Loop, & Analytics
The collection areas will now add an item to the player's inventory. The player character will move to a collection area if clicked on. Score is updated upon collecting an item, but this will be later changed to completion of a list. Game now ends when timer reaches 0. Analytics have been moved to the end of a round.